feat: move inner function with mutation to separate files
