+++ packages/cli/src/commands/join/helpers/index.ts
@@ -10,3 +10,9 @@ export * from './show-conflicts.js';
 export * from './prefix-tag-suggestion.js';
 export * from './duplicate-tag-description-warning.js';
 export * from './iterate-potential-conflicts.js';
+export * from './populate-tags.js';
+export * from './collect-external-docs.js';
+export * from './collect-paths.js';
+export * from './collect-components.js';
+export * from './collect-webhooks.js';
+export * from './add-info-section.js';
