+++ packages/cli/src/commands/join/index.ts
@@ -4,26 +4,12 @@ import {
   getTotals,
   detectSpec,
   bundleDocument,
-  isRef,
-  dequal,
   logger,
   isPlainObject,
-  keysOf,
   isEmptyObject,
   getTypes,
 } from '@redocly/openapi-core';
-import type {
-  Document,
-  Referenced,
-  Exact,
-  BundleResult,
-  Oas3PathItem,
-  Oas3Server,
-  Oas3Tag,
-  Oas3_2Tag,
-  SpecVersion,
-  Oas3Parameter,
-} from '@redocly/openapi-core';
+import type { Document, BundleResult, Oas3Server, SpecVersion } from '@redocly/openapi-core';
 import { blue, yellow } from 'colorette';
 import * as path from 'node:path';
 import { performance } from 'node:perf_hooks';
@@ -37,21 +23,20 @@ import {
   writeToFileByExtension,
 } from '../../utils/miscellaneous.js';
 import type { CommandArgs } from '../../wrapper.js';
-import { COMPONENTS, type Oas3Method, OPENAPI3_METHOD_NAMES } from '../split/types.js';
+import { COMPONENTS } from '../split/types.js';
 import {
   replace$Refs,
   getInfoPrefix,
-  addSecurityPrefix,
-  addPrefix,
-  addComponentsPrefix,
-  formatTags,
   getApiFilename,
   iteratePotentialConflicts,
+  populateTags,
+  collectExternalDocs,
+  collectPaths,
+  collectComponents,
+  collectWebhooks,
+  addInfoSectionAndSpecVersion,
 } from './helpers/index.js';
-import type { JoinArgv, JoinDocumentContext, AnyOas3Definition } from './types.js';
-
-const Tags = 'tags';
-const xTagGroups = 'x-tagGroups';
+import type { JoinArgv, AnyOas3Definition } from './types.js';
 
 export async function handleJoin({
   argv,
@@ -180,7 +165,7 @@ export async function handleJoin({
     webhooks: {},
   };
 
-  addInfoSectionAndSpecVersion(documents, prefixComponentsWithInfoProp);
+  addInfoSectionAndSpecVersion(joinedDef, documents, prefixComponentsWithInfoProp);
 
   if (serversAreTheSame && first.parsed.servers) {
     joinedDef.servers = first.parsed.servers;
@@ -213,12 +198,12 @@ export async function handleJoin({
       oasVersion,
     };
     if (tags) {
-      populateTags(context);
+      populateTags({ joinedDef, withoutXTagGroups, context });
     }
-    collectExternalDocs(openapi, context);
-    collectPaths(openapi, context, serversAreTheSame);
-    collectComponents(openapi, context);
-    collectWebhooks(openapi, context);
+    collectExternalDocs({ joinedDef, openapi, context });
+    collectPaths({ joinedDef, withoutXTagGroups, openapi, context, serversAreTheSame });
+    collectComponents({ joinedDef, openapi, context });
+    collectWebhooks({ joinedDef, withoutXTagGroups, openapi, context });
     if (componentsPrefix) {
       replace$Refs(openapi, componentsPrefix);
     }
@@ -237,417 +222,4 @@ export async function handleJoin({
   writeToFileByExtension(sortTopLevelKeysForOas(joinedDef), specFilename, noRefs);
 
   printExecutionTime('join', startedAt, specFilename);
-
-  function populateTags({
-    api,
-    apiFilename,
-    apiTitle,
-    tags,
-    potentialConflicts,
-    tagsPrefix,
-    componentsPrefix,
-    oasVersion,
-  }: JoinDocumentContext) {
-    if (!joinedDef.hasOwnProperty(Tags)) {
-      joinedDef[Tags] = [];
-    }
-    if (!potentialConflicts.tags.hasOwnProperty('all')) {
-      potentialConflicts.tags['all'] = {};
-    }
-    if (withoutXTagGroups && !potentialConflicts.tags.hasOwnProperty('description')) {
-      potentialConflicts.tags['description'] = {};
-    }
-    for (const tag of tags || []) {
-      const entrypointTagName = addPrefix(tag.name, tagsPrefix);
-      if (tag.description) {
-        tag.description = addComponentsPrefix(tag.description, componentsPrefix!);
-      }
-
-      const tagDuplicate = joinedDef.tags.find(
-        (t: Oas3Tag | Oas3_2Tag) => t.name === entrypointTagName
-      );
-
-      if (tagDuplicate && withoutXTagGroups) {
-        // If tag already exist and `without-x-tag-groups` option,
-        // check if description are different for potential conflicts warning.
-        const isTagDescriptionNotEqual =
-          tag.hasOwnProperty('description') && tagDuplicate.description !== tag.description;
-
-        potentialConflicts.tags.description[entrypointTagName].push(
-          ...(isTagDescriptionNotEqual ? [api] : [])
-        );
-      } else if (!tagDuplicate) {
-        // Instead add tag to joinedDef if there no duplicate;
-        if (oasVersion === 'oas3_0' || oasVersion === 'oas3_1') {
-          (tag as Oas3Tag)['x-displayName'] = (tag as Oas3Tag)['x-displayName'] || tag.name;
-        } else if (oasVersion === 'oas3_2') {
-          (tag as Oas3_2Tag).summary = (tag as Oas3_2Tag).summary || tag.name;
-        }
-        tag.name = entrypointTagName;
-        joinedDef.tags.push(tag);
-
-        if (withoutXTagGroups) {
-          potentialConflicts.tags.description[entrypointTagName] = [api];
-        }
-      }
-
-      if (!withoutXTagGroups && oasVersion !== 'oas3_2') {
-        const groupName = apiTitle || apiFilename;
-        createXTagGroups(groupName);
-        if (!tagDuplicate) {
-          populateXTagGroups(entrypointTagName, getIndexGroup(groupName));
-        }
-      }
-
-      const doesEntrypointExist =
-        !potentialConflicts.tags.all[entrypointTagName] ||
-        (potentialConflicts.tags.all[entrypointTagName] &&
-          !potentialConflicts.tags.all[entrypointTagName].includes(api));
-      potentialConflicts.tags.all[entrypointTagName] = [
-        ...(potentialConflicts.tags.all[entrypointTagName] || []),
-        ...(!withoutXTagGroups && doesEntrypointExist ? [api] : []),
-      ];
-    }
-  }
-
-  function getIndexGroup(name: string): number {
-    return joinedDef[xTagGroups].findIndex((item: any) => item.name === name);
-  }
-
-  function createXTagGroups(name: string) {
-    if (!joinedDef.hasOwnProperty(xTagGroups)) {
-      joinedDef[xTagGroups] = [];
-    }
-
-    if (!joinedDef[xTagGroups].some((g: any) => g.name === name)) {
-      joinedDef[xTagGroups].push({ name, tags: [] });
-    }
-
-    const indexGroup = getIndexGroup(name);
-
-    if (!joinedDef[xTagGroups][indexGroup].hasOwnProperty(Tags)) {
-      joinedDef[xTagGroups][indexGroup][Tags] = [];
-    }
-  }
-
-  function populateXTagGroups(entrypointTagName: string, indexGroup: number) {
-    if (
-      !joinedDef[xTagGroups][indexGroup][Tags].find((t: Oas3Tag) => t.name === entrypointTagName)
-    ) {
-      joinedDef[xTagGroups][indexGroup][Tags].push(entrypointTagName);
-    }
-  }
-
-  function collectExternalDocs(openapi: AnyOas3Definition, { api }: JoinDocumentContext) {
-    const { externalDocs } = openapi;
-    if (externalDocs) {
-      if (joinedDef.hasOwnProperty('externalDocs')) {
-        logger.warn(`warning: skip externalDocs from ${blue(path.basename(api))} \n`);
-        return;
-      }
-      joinedDef['externalDocs'] = externalDocs;
-    }
-  }
-
-  function collectPaths(
-    openapi: AnyOas3Definition,
-    {
-      apiFilename,
-      apiTitle,
-      api,
-      potentialConflicts,
-      tagsPrefix,
-      componentsPrefix,
-      oasVersion,
-    }: JoinDocumentContext,
-    serversAreTheSame: boolean
-  ) {
-    const { paths, servers: rootServers } = openapi;
-    const operationsSet = new Set(OPENAPI3_METHOD_NAMES);
-    if (paths) {
-      if (!joinedDef.hasOwnProperty('paths')) {
-        joinedDef['paths'] = {};
-      }
-
-      for (const path of keysOf(paths)) {
-        if (!joinedDef.paths.hasOwnProperty(path)) {
-          joinedDef.paths[path] = {};
-        }
-        if (!potentialConflicts.paths.hasOwnProperty(path)) {
-          potentialConflicts.paths[path] = {};
-        }
-
-        const pathItem = paths[path] as Oas3PathItem;
-        const servers = serversAreTheSame
-          ? pathItem.servers
-          : pathItem.servers || rootServers || [];
-        if (servers) {
-          collectPathServers(servers, path);
-        }
-
-        for (const field of keysOf(pathItem)) {
-          if (operationsSet.has(field as Oas3Method)) {
-            collectPathOperation(pathItem, path, field as Oas3Method);
-          }
-          if (field === 'parameters') {
-            collectPathParameters(pathItem, path);
-          }
-          if (typeof pathItem[field] === 'string') {
-            collectPathStringFields(pathItem, path, field);
-          }
-        }
-      }
-    }
-
-    function collectPathStringFields(
-      pathItem: Oas3PathItem,
-      path: string | number,
-      field: keyof Oas3PathItem
-    ) {
-      const fieldValue = pathItem[field];
-      if (
-        joinedDef.paths[path].hasOwnProperty(field) &&
-        joinedDef.paths[path][field] !== fieldValue
-      ) {
-        logger.warn(`warning: different ${field} values in ${path}\n`);
-        return;
-      }
-      joinedDef.paths[path][field] = fieldValue;
-    }
-
-    function collectPathServers(servers: Oas3Server[], path: string | number) {
-      if (!servers) {
-        return;
-      }
-
-      if (!joinedDef.paths[path].hasOwnProperty('servers')) {
-        joinedDef.paths[path].servers = [];
-      }
-
-      for (const server of servers) {
-        let isFoundServer = false;
-        for (const pathServer of joinedDef.paths[path].servers) {
-          if (pathServer.url === server.url) {
-            if (!isServersEqual(pathServer, server)) {
-              exitWithError(`Different server values for (${server.url}) in ${path}.`);
-            }
-            isFoundServer = true;
-          }
-        }
-
-        if (!isFoundServer) {
-          joinedDef.paths[path].servers.push(server);
-        }
-      }
-    }
-
-    function collectPathParameters(pathItem: Oas3PathItem, path: string | number) {
-      if (!pathItem.parameters) {
-        return;
-      }
-      if (!joinedDef.paths[path].hasOwnProperty('parameters')) {
-        joinedDef.paths[path].parameters = [];
-      }
-
-      for (const parameter of pathItem.parameters as Referenced<Oas3Parameter>[]) {
-        let isFoundParameter = false;
-
-        for (const pathParameter of joinedDef.paths[path]
-          .parameters as Referenced<Oas3Parameter>[]) {
-          // Compare $ref only if both are reference objects
-          if (isRef(pathParameter) && isRef(parameter)) {
-            if (pathParameter['$ref'] === parameter['$ref']) {
-              isFoundParameter = true;
-            }
-          }
-          // Compare properties only if both are reference objects
-          if (!isRef(pathParameter) && !isRef(parameter)) {
-            if (pathParameter.name === parameter.name && pathParameter.in === parameter.in) {
-              if (!dequal(pathParameter.schema, parameter.schema)) {
-                exitWithError(`Different parameter schemas for (${parameter.name}) in ${path}.`);
-              }
-              isFoundParameter = true;
-            }
-          }
-        }
-
-        if (!isFoundParameter) {
-          joinedDef.paths[path].parameters.push(parameter);
-        }
-      }
-    }
-
-    function collectPathOperation(
-      pathItem: Oas3PathItem,
-      path: string | number,
-      operation: Oas3Method
-    ) {
-      const pathOperation = pathItem[operation];
-
-      if (!pathOperation) {
-        return;
-      }
-
-      joinedDef.paths[path][operation] = pathOperation;
-      potentialConflicts.paths[path][operation] = [
-        ...(potentialConflicts.paths[path][operation] || []),
-        api,
-      ];
-
-      const { operationId } = pathOperation;
-
-      if (operationId) {
-        if (!potentialConflicts.paths.hasOwnProperty('operationIds')) {
-          potentialConflicts.paths['operationIds'] = {};
-        }
-        potentialConflicts.paths.operationIds[operationId] = [
-          ...(potentialConflicts.paths.operationIds[operationId] || []),
-          api,
-        ];
-      }
-
-      const { tags, security } = joinedDef.paths[path][operation];
-
-      if (tags) {
-        joinedDef.paths[path][operation].tags = tags.map((tag: string) =>
-          addPrefix(tag, tagsPrefix)
-        );
-        populateTags({
-          api,
-          apiFilename,
-          apiTitle,
-          tags: formatTags(tags),
-          potentialConflicts,
-          tagsPrefix,
-          componentsPrefix,
-          oasVersion,
-        });
-      } else {
-        joinedDef.paths[path][operation]['tags'] = [addPrefix('other', tagsPrefix || apiFilename)];
-        populateTags({
-          api,
-          apiFilename,
-          apiTitle,
-          tags: formatTags(['other']),
-          potentialConflicts,
-          tagsPrefix: tagsPrefix || apiFilename,
-          componentsPrefix,
-          oasVersion,
-        });
-      }
-      if (!security && openapi.hasOwnProperty('security')) {
-        joinedDef.paths[path][operation]['security'] = addSecurityPrefix(
-          openapi.security,
-          componentsPrefix!
-        );
-      } else if (pathOperation.security) {
-        joinedDef.paths[path][operation].security = addSecurityPrefix(
-          pathOperation.security,
-          componentsPrefix!
-        );
-      }
-    }
-  }
-
-  function isServersEqual(serverOne: Oas3Server, serverTwo: Oas3Server) {
-    if (serverOne.description === serverTwo.description) {
-      return dequal(serverOne.variables, serverTwo.variables);
-    }
-
-    return false;
-  }
-
-  function collectComponents(
-    openapi: AnyOas3Definition,
-    { api, potentialConflicts, componentsPrefix }: JoinDocumentContext
-  ) {
-    const { components } = openapi;
-    if (components) {
-      if (!joinedDef.hasOwnProperty(COMPONENTS)) {
-        joinedDef[COMPONENTS] = {};
-      }
-      for (const [component, componentObj] of Object.entries(components)) {
-        if (!potentialConflicts[COMPONENTS].hasOwnProperty(component)) {
-          potentialConflicts[COMPONENTS][component] = {};
-          joinedDef[COMPONENTS][component] = {};
-        }
-        for (const item of Object.keys(componentObj)) {
-          const componentPrefix = addPrefix(item, componentsPrefix!);
-          potentialConflicts.components[component][componentPrefix] = [
-            ...(potentialConflicts.components[component][item] || []),
-            { [api]: componentObj[item] },
-          ];
-          joinedDef.components[component][componentPrefix] = componentObj[item];
-        }
-      }
-    }
-  }
-
-  function collectWebhooks(
-    openapi: AnyOas3Definition,
-    {
-      apiFilename,
-      apiTitle,
-      api,
-      potentialConflicts,
-      tagsPrefix,
-      componentsPrefix,
-      oasVersion,
-    }: JoinDocumentContext
-  ) {
-    const webhooks = oasVersion === 'oas3_0' ? 'x-webhooks' : 'webhooks';
-    const openapiWebhooks = (openapi as Exact<AnyOas3Definition>)[webhooks];
-    if (openapiWebhooks) {
-      if (!joinedDef.hasOwnProperty(webhooks)) {
-        joinedDef[webhooks] = {};
-      }
-      for (const webhook of Object.keys(openapiWebhooks)) {
-        joinedDef[webhooks][webhook] = openapiWebhooks[webhook];
-
-        if (!potentialConflicts.webhooks.hasOwnProperty(webhook)) {
-          potentialConflicts.webhooks[webhook] = {};
-        }
-        for (const operation of Object.keys(openapiWebhooks[webhook])) {
-          potentialConflicts.webhooks[webhook][operation] = [
-            ...(potentialConflicts.webhooks[webhook][operation] || []),
-            api,
-          ];
-        }
-        for (const operationKey of Object.keys(joinedDef[webhooks][webhook])) {
-          const { tags } = joinedDef[webhooks][webhook][operationKey];
-          if (tags) {
-            joinedDef[webhooks][webhook][operationKey].tags = tags.map((tag: string) =>
-              addPrefix(tag, tagsPrefix)
-            );
-            populateTags({
-              api,
-              apiFilename,
-              apiTitle,
-              tags: formatTags(tags),
-              potentialConflicts,
-              tagsPrefix,
-              componentsPrefix,
-              oasVersion,
-            });
-          }
-        }
-      }
-    }
-  }
-
-  function addInfoSectionAndSpecVersion(
-    documents: any,
-    prefixComponentsWithInfoProp: string | undefined
-  ) {
-    const firstApi = documents[0];
-    const openapi = firstApi.parsed;
-    const componentsPrefix = getInfoPrefix(openapi.info, prefixComponentsWithInfoProp, COMPONENTS);
-    if (!openapi.openapi) exitWithError('Version of specification is not found.');
-    if (!openapi.info) exitWithError('Info section is not found in specification.');
-    if (openapi.info?.description) {
-      openapi.info.description = addComponentsPrefix(openapi.info.description, componentsPrefix);
-    }
-    joinedDef.openapi = openapi.openapi;
-    joinedDef.info = openapi.info;
-  }
 }
